move all sample obfuscator configs beside their classes
