+++ src/churn/churn.ts
@@ -7,15 +7,12 @@ import caesar = require('../simple-transformers/caesar');
 import candidate = require('./candidate');
 import churn_pipe_types = require('../churn-pipe/freedom-module.interface');
 import churn_types = require('./churn.types');
-import encryption = require('../fancy-transformers/encryptionShaper');
 import handler = require('../handler/queue');
 import ipaddr = require('ipaddr.js');
 import logging = require('../logging/logging');
 import net = require('../net/net.types');
 import peerconnection = require('../webrtc/peerconnection');
-import protean = require('../fancy-transformers/protean');
 import random = require('../crypto/random');
-import sequence = require('../fancy-transformers/byteSequenceShaper');
 import signals = require('../webrtc/signals');
 
 import ChurnSignallingMessage = churn_types.ChurnSignallingMessage;
@@ -290,34 +287,6 @@ export var filterCandidatesFromSdp = (sdp:string) : string => {
       this.pipe_.setTransformer(obfuscatorConfig);
     }
 
-    private makeSampleEncryptionConfig_ = () :encryption.EncryptionConfig => {
-      var key = new ArrayBuffer(16);
-      return {key: arraybuffers.arrayBufferToHexString(key)};
-    }
-
-    private makeSampleSequences_ = () :sequence.SequenceConfig => {
-      var buffer = arraybuffers.stringToArrayBuffer("OH HELLO");
-      var hex = arraybuffers.arrayBufferToHexString(buffer);
-      var sequence = {
-        index: 0,
-        offset: 0,
-        sequence: hex,
-        length: 256
-      };
-
-      return {
-        addSequences: [sequence],
-        removeSequences: [sequence]
-      };
-    }
-
-    private makeSampleProteanConfig_ = () :protean.ProteanConfig => {
-      return {
-        encryption: this.makeSampleEncryptionConfig_(),
-        injection: this.makeSampleSequences_()
-      };
-    }
-
     private addRemoteCandidate_ = (iceCandidate:RTCIceCandidate) => {
       var c = Candidate.fromRTCIceCandidate(iceCandidate);
       var remoteEndpoint = {
@@ -430,7 +399,7 @@ export var filterCandidatesFromSdp = (sdp:string) : string => {
         });
         this.haveObfuscatorConfig_(this.preferredObfuscatorConfig_);
       } else {
-        var caesarConfig = caesar.makeRandomConfig();
+        var caesarConfig = caesar.sampleConfig();
         this.signalForPeerQueue.handle({
           caesar: caesarConfig.key
         });
